# apps/backend/src/routes/visitors.ts
import { Router, type IRouter, type Request, type Response } from "express";
import { db, visitorsTable, visitsTable, sectorsTable, usersTable } from "@visit-control/db";
import { eq, ilike, or, and, desc, sql, type SQL } from "drizzle-orm";
import { requireAuth, requireAdmin } from "../middlewares/auth";
import { auditAction } from "../lib/audit";

type AuthReq = Request & { user: typeof usersTable.$inferSelect };

const router: IRouter = Router();

function formatVisitor(v: typeof visitorsTable.$inferSelect) {
  return {
    ...v,
    createdAt: v.createdAt.toISOString(),
    updatedAt: v.updatedAt?.toISOString() ?? null,
  };
}

router.get("/visitors/search", requireAuth, async (req: Request, res: Response): Promise<void> => {
  const q = req.query.q as string | undefined;
  if (!q || q.trim().length < 2) {
    res.json([]);
    return;
  }

  const visitors = await db
    .select()
    .from(visitorsTable)
    .where(
      or(
        ilike(visitorsTable.name, `%${q}%`),
        ilike(visitorsTable.cpf, `%${q}%`),
      ),
    )
    .orderBy(visitorsTable.name)
    .limit(10);

  res.json(visitors.map(formatVisitor));
});

router.get("/visitors", requireAuth, async (req: Request, res: Response): Promise<void> => {
  const { search, cpf, phone, company, city, page = "1", limit = "20" } =
    req.query as Record<string, string | undefined>;

  const pageNum = Math.max(1, parseInt(page, 10));
  const limitNum = Math.min(100, parseInt(limit, 10));
  const offset = (pageNum - 1) * limitNum;

  const conditions: SQL[] = [];
  if (search) conditions.push(ilike(visitorsTable.name, `%${search}%`));
  if (cpf) conditions.push(ilike(visitorsTable.cpf, `%${cpf}%`));
  if (phone) conditions.push(ilike(visitorsTable.phone, `%${phone}%`));
  if (company) conditions.push(ilike(visitorsTable.company, `%${company}%`));
  if (city) conditions.push(ilike(visitorsTable.city, `%${city}%`));

  const whereClause = conditions.length ? and(...conditions) : undefined;

  const [{ count }] = await db
    .select({ count: sql<number>`count(*)::int` })
    .from(visitorsTable)
    .where(whereClause);

  const visitors = await db
    .select()
    .from(visitorsTable)
    .where(whereClause)
    .orderBy(visitorsTable.name)
    .limit(limitNum)
    .offset(offset);

  res.json({
    data: visitors.map(formatVisitor),
    total: count,
    page: pageNum,
    limit: limitNum,
  });
});

router.post("/visitors", requireAuth, async (req: Request, res: Response): Promise<void> => {
  const { name, cpf, phone, company, city } = req.body ?? {};
  if (!name) {
    res.status(400).json({ error: "Nome é obrigatório" });
    return;
  }

  const [visitor] = await db
    .insert(visitorsTable)
    .values({
      name: String(name),
      cpf: cpf ? String(cpf) : null,
      phone: phone ? String(phone) : null,
      company: company ? String(company) : null,
      city: city ? String(city) : null,
    })
    .returning();

  await auditAction({
    userId: (req as AuthReq).user.id,
    action: "create_visitor",
    ipAddress: req.ip,
    entityType: "visitor",
    entityId: visitor.id,
    newData: { name, cpf, phone, company, city },
  });

  res.status(201).json(formatVisitor(visitor));
});

router.get("/visitors/:id", requireAuth, async (req: Request, res: Response): Promise<void> => {
  const id = parseInt(Array.isArray(req.params.id) ? req.params.id[0] : req.params.id, 10);
  const [visitor] = await db.select().from(visitorsTable).where(eq(visitorsTable.id, id));
  if (!visitor) {
    res.status(404).json({ error: "Visitante não encontrado" });
    return;
  }

  const visits = await db
    .select({
      id: visitsTable.id,
      visitorId: visitsTable.visitorId,
      sectorId: visitsTable.sectorId,
      sector: {
        id: sectorsTable.id,
        name: sectorsTable.name,
        abbreviation: sectorsTable.abbreviation,
        secretariat: sectorsTable.secretariat,
        status: sectorsTable.status,
        createdAt: sectorsTable.createdAt,
      },
      responsible: visitsTable.responsible,
      reason: visitsTable.reason,
      notes: visitsTable.notes,
      status: visitsTable.status,
      entryDate: visitsTable.entryDate,
      entryTime: visitsTable.entryTime,
      entryUserId: visitsTable.entryUserId,
      exitDate: visitsTable.exitDate,
      exitTime: visitsTable.exitTime,
      exitUserId: visitsTable.exitUserId,
      cancelReason: visitsTable.cancelReason,
      createdAt: visitsTable.createdAt,
    })
    .from(visitsTable)
    .leftJoin(sectorsTable, eq(visitsTable.sectorId, sectorsTable.id))
    .where(eq(visitsTable.visitorId, id))
    .orderBy(desc(visitsTable.createdAt));

  res.json({
    ...formatVisitor(visitor),
    visits: visits.map((v) => ({
      ...v,
      createdAt: v.createdAt.toISOString(),
      sector: v.sector
        ? { ...v.sector, createdAt: v.sector.createdAt.toISOString() }
        : null,
    })),
  });
});

router.patch("/visitors/:id", requireAuth, requireAdmin, async (req: Request, res: Response): Promise<void> => {
  const id = parseInt(Array.isArray(req.params.id) ? req.params.id[0] : req.params.id, 10);
  const [existing] = await db.select().from(visitorsTable).where(eq(visitorsTable.id, id));
  if (!existing) {
    res.status(404).json({ error: "Visitante não encontrado" });
    return;
  }

  const { name, cpf, phone, company, city } = req.body ?? {};
  const updates: Partial<typeof visitorsTable.$inferInsert> = { updatedAt: new Date() };
  if (name) updates.name = String(name);
  if (cpf !== undefined) updates.cpf = cpf ? String(cpf) : null;
  if (phone !== undefined) updates.phone = phone ? String(phone) : null;
  if (company !== undefined) updates.company = company ? String(company) : null;
  if (city !== undefined) updates.city = city ? String(city) : null;

  const [updated] = await db
    .update(visitorsTable)
    .set(updates)
    .where(eq(visitorsTable.id, id))
    .returning();

  await auditAction({
    userId: (req as AuthReq).user.id,
    action: "update_visitor",
    ipAddress: req.ip,
    entityType: "visitor",
    entityId: id,
    previousData: formatVisitor(existing),
    newData: req.body,
  });

  res.json(formatVisitor(updated));
});

export default router;