# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Sistema de Controle de Visitantes
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /visitors:
    get:
      operationId: listVisitors
      tags: [visitors]
      summary: List visitors with optional search
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: cpf
          schema:
            type: string
        - in: query
          name: phone
          schema:
            type: string
        - in: query
          name: company
          schema:
            type: string
        - in: query
          name: city
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of visitors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitorListResponse"
    post:
      operationId: createVisitor
      tags: [visitors]
      summary: Create visitor
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visitor"
  /visitors/search:
    get:
      operationId: searchVisitors
      tags: [visitors]
      summary: Real-time search for autocomplete during visit registration
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matched visitors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Visitor"
  /visitors/{id}:
    get:
      operationId: getVisitor
      tags: [visitors]
      summary: Get visitor with full visit history
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Visitor with history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VisitorWithHistory"
    patch:
      operationId: updateVisitor
      tags: [visitors]
      summary: Update visitor
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Visitor"

  # ── Visits ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: integer
  schemas:
...
    User:
      type: object
      required: [id, name, login, role, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        login:
          type: string
        role:
          type: string
          enum: [admin, receptionist]
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
...
    Sector:
      type: object
      required: [id, name, abbreviation, secretariat, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        abbreviation:
          type: string
        secretariat:
          type: string
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
...
    Visitor:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        cpf:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
    VisitorInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        cpf:
          type: string
        phone:
          type: string
        company:
          type: string
        city:
          type: string
    VisitorUpdate:
      type: object
      properties:
        name:
          type: string
        cpf:
          type: string
        phone:
          type: string
        company:
          type: string
        city:
          type: string
    VisitorWithHistory:
      type: object
      required: [id, name, createdAt, visits]
      properties:
        id:
          type: integer
        name:
          type: string
        cpf:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
        visits:
          type: array
          items:
            $ref: "#/components/schemas/Visit"
    VisitorListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Visitor"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── Visit ─────────────────────────────────────────────────────────────
    Visit:
      type: object
      required: [id, visitorId, sectorId, status, entryDate, entryTime, entryUserId, createdAt]
      properties:
        id:
          type: integer
        visitorId:
          type: integer
        visitor:
          $ref: "#/components/schemas/Visitor"
        sectorId:
          type: integer
        sector:
          $ref: "#/components/schemas/Sector"
        responsible:
          type: ["string", "null"]
        reason:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [ongoing, finished, cancelled]
        entryDate:
          type: string
          format: date
        entryTime:
          type: string
        entryUserId:
          type: integer
        entryUser:
          $ref: "#/components/schemas/User"
        exitDate:
          type: ["string", "null"]
          format: date
        exitTime:
          type: ["string", "null"]
        exitUserId:
          type: ["integer", "null"]
        exitUser:
          $ref: "#/components/schemas/User"
        cancelReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...