# packages/api-zod/src/generated/types/visitorWithHistory.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Sistema de Controle de Visitantes
 * OpenAPI spec version: 0.1.0
 */
import type { Visit } from './visit';

export interface VisitorWithHistory {
  id: number;
  name: string;
  /** @nullable */
  cpf?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  company?: string | null;
  /** @nullable */
  city?: string | null;
  createdAt: Date;
  /** @nullable */
  updatedAt?: Date | null;
  visits: Visit[];
}